Separate listeners into a extra class from module main class.
